[gnugo] New command: ‘S’ (gnugo-request-suggestion)
* packages/gnugo/gnugo.el (gnugo-gate):
If waiting for a suggestion, say "Still thinking"
instead of "Not your turn yet" in error message.
(gnugo--q): If waiting for a suggestion, say
"receive a suggestion" instead of "play" in error message.
(gnugo--rename-buffer-portion): New func.
(gnugo-get-move-insertion-filter): On received suggestion,
rename buffer w/ "to play", warp the cursor unless nonsensical
or inhibited, and display the suggestion in the echo area.
(gnugo-get-move): Take optional arg SUGGESTION;
save it as well as color in :waiting.
(gnugo-request-suggestion): New command.
(gnugo-board-mode-map): Add binding for ‘S’.