Don't let el-search-pattern resume a paused search
When calling `el-search-pattern' from the head buffer of the
el-search--current-search, don't automatically resume the
el-search--current-search when the last command was not a search
command even when the search pattern is the same.  Start a new
single-buffer search as normal.  `el-search-jump-to-search-head' is
now the preferred way to resume a suspended el-search--current-search.