* psgml-parse.el (sgml-pi-asl-parser): New function
Copyright-paperwork-exempt: yes

A convenience function for parsing the text of a PI like an attribute
specification list with a couple syntax enhancements.  May be called to
parse the PI instead of using the string argument to `sgml-pi-function'.

The makes available for other processing the parsing work of
`sgml--pi-element-handler'.

(sgml-parse-name-or-literal, sgml-pi-asl-parser): New functions extracted from
 sgml--pi-element-handler.
(sgml--pi-element-handler): Reimplement using those.