Improve working of `el-search-kill-left-over-search-buffers'
We used to consult the ring elements of el-search-history to identify
buffers with non-nil el-search--temp-buffer-flag which might still be
relevant to the user, which was not optimal.  With this change,
`el-search--pending-search-p' becomes a reliable mean to do that, and
we can reimplement el-search-kill-left-over-search-buffers to use this
function instead.

* el-search/el-search.el (el-search--next-buffer): Remove
`el-search-hl-post-command-fun' from the local binding of
`post-command-hook' in the old buffer so that
`el-search--pending-search-p' is guaranteed to return nil there.
(el-search-kill-left-over-search-buffers): Use
`el-search--pending-search-p' to identify temporarily opened file
buffers which we should not kill.