[el-search] Make replacement editable and ediff'able
Make the 'el-search-query-replace' replacement editable and ediff'able
using a temporary buffer.  Update header comment accordingly.

* packages/el-search/el-search.el (el-search--replace-hunk): Return a
marker pointing to the end of the inserted text.
(el-search-query-replace-ediff-replacement): New function.
(el-search--search-and-replace-pattern): Add and implement keys to
edit the replacement in a separate buffer and to ediff the current
match with the current replacement.