[el-search] Fix behavior after completed single-buffer search
Make 'el-search-from-beginning' and 'el-search-jump-to-search-head'
after completed single-buffer search work as expected.

* packages/el-search/el-search.el (el-search-pattern)
(el-search-pattern-backward, el-search-this-sexp)
(el-search--occur-button-action)
(el-search--search-and-replace-pattern): Make single-buffer ad-hoc
searches have the search buffer as non-nil 'is-single-buffer' value.
(el-search-object): Add comment about the intended usage of the
'is-single-buffer' property.
(el-search-revive-search): New function.
(el-search-barf-if-not-search-buffer): If we are in the searched
buffer of a completed single-buffer search, silently revive it with
'el-search-revive-search'.
(el-search-jump-to-search-head): When the search needs to be
restarted, use 'el-search-revive-search' instead of only
'el-search-reset-search'.