[el-search] Minibuffer pattern prompt hints part 3
Restore the ability of 'el-search-read-display-mb-hints' to warn about
errors happening while counting (i.e. searching for) matches, i.e. for
patterns that are syntactically valid and compile ok but raise an
error when applied to some value.  That capability had been gotten
lost with the fix in "Fix quits in minibuffer in emacs -nw".
We also let 'el-search-display-match-count' accept an ANIMATOR
function argument.

* packages/el-search/el-search.el (el-search-display-match-count):
Allow to pass an animator function as second argument.  Change callers
to do that.
(el-search-read-display-mb-hints): Handle cases when
'el-search-display-match-count' exits nonlocally.