* smalltalk-mode.el: Heed compiler warnings; minor improvements
Set lexical-binding to t: The code should work with both settings (so it still
works on older Emacsen), but we get better warnings this way.
Improve Commentary section.  Fix broken "A-z" character set in regexps.
Fix broken highlighting when `$` is the last char in a string.
Fix compiler warnings, such as removing unused vars.
Use syntax-ppss instead of parse-partial-sexp.

(smalltalk-mode): New Customize group.
(smalltalk-whitespace): Make it a defcustom.
(smalltalk-mode-abbrev-table): Merge declaration and definition.
(smalltalk-mode-map): Remove LF binding which is a user-preference
rather than something related to the buffer's syntax.
(smalltalk--when-fboundp): New macro.
(smalltalk--syntax-propertize): New var.
(smalltalk-mode): Use it.  Use define-derived-mode.  Set comment-end-skip.
(smalltalk-bang-begin-of-defun): Use (nth 8 (syntax-ppss)) to jump
directly to the beginning of a string.
(smalltalk-reindent): Make it an alias of indent-according-to-mode.
(smalltalk-newline-and-indent): Make it an alias of newline-and-indent.
(smalltalk-bang): Use (nth 8 (syntax-ppss)).
(smalltalk-comment-indent): Let newcomment.el deal with current-column.
(smalltalk-toplevel-indent, smalltalk-at-begin-of-scope)
(smalltalk-at-begin-of-defun):
Avoid line-number-at-pos, which is slow in large buffers.
(smalltalk-in-string): Delete function.
(smalltalk-in-comment): Simplify.
(smalltalk-parse-sexp-and-narrow-to-paren): Use syntax-ppss.
(smalltalk-narrow-to-method): Don't burp when hitting EOB.