[el-search] Extend possible search domains
Allow search entities that are neither live buffers nor files.  This
makes it possible to implement el-search commands that e.g. search old
revisions of a repository.

(el-search--get-buffer-fun): New buffer local helper var.
(el-search--occur): Use value of 'el-search--get-buffer-fun' when
bound in any buffer in the matches stream to enable
'el-search--occur-button-action' to recreate that buffer.
(el-search--occur-button-action): Allow the first argument to be a
function to create the match buffer.