etc/poke-mode.el: add Emacs mode for .pk files
Add simple Emacs mode for Poke programs.
- Written from "scratch" as cc-mode is insane
- Highlights
  * comments, shebang, string, char
  * keywords
  * builtins (functions, vars, exceptions)
  * attributes (expr'attr)
  * unit suffixes (expr#unit)
- Customizable faces for each highlighted things
- No indentation yet

There are some issues as Poke allows some identifiers to use language
keywords e.g. "offset" is lexed as a token for the offset type but can
also be used as a var name. Poke-mode highlights both cases as a type.

Trying it out
=============

Add the following to your Emacs init file:

  (load-file "path/to/poke-mode.el)
  (add-to-list 'auto-mode-alist '("\\.pk\\'" . poke-mode))

Customizing the faces
=====================

To change the face used for a piece of text, place the point on it and
run M-x customize-face RET RET.

2020-06-08  Aurélien Aptel  <aaptel@suse.com>

	* etc/poke-mode.el: New file.
	* HACKING (Writing Poke): Mention poke-mode.