Loop fix and autoload cookie, docstring clarifications
* rec-mode.el (rec-record-descriptor): Use cl-loop to come up with something
that isn't O(n^2) by creating a list of "next" descriptors. The main loop body
is the same: the `return' form is just a convenient way of doing `(setq found
t)'.  It's probably not "better" (in terms of readability, etc.) but cl-loop
should generate something quite efficient.
(rec-cmd-goto-next-rec): Clarify it moves to the record of the
same type.
(rec-cmd-goto-previous-rec): Ditto.