Move smartly around after editing fields
* rec-mode.el (rec-edit-buffer): Recenter after viewing the whole
buffer. Useful if we're at the last record, we might want to see
the rest of the buffer.
(rec-cmd-edit-field): Move to next field after editing in the
various special modes.
(rec-move-to-next-line-after-edit): Add. Controls the above.
(rec-finish-editing-move): Add. Implements the functionality.
(rec-finish-editing-field): Restore previous window configuration,
and move if we were in record edition mode.
(rec-editing): Add docstring and FIXME.