add loading functionality
* blist.el (blist-import-new-list): A modified loading function.  It
uses a hash table to test if new names are found in the old
bookmark-alist, so that the overall worst case time complexity is
linear, rather than quadratic.  Also the behaviour of `bookmark-load'
differs from that described by its documentation string: it puts the
new list at the end of the list in fact.  That is fixed in this
function as well.

(blist-load): Adapted version of `bookmark-bmenu-load'.

(blist-delete-marked, blist-do-delete): Increase
bookmark-alist-modification-count so that we can correctly
automatically save bookmarks if the users demands so.