ilist: leave new lines when hiding groups
* ilist.el (ilist-length, ilist-display): Replace ilist-length by
  string-width.

(ilist-string): Leave a newline at the end of a group without the
invisible property.  Then hiding the group will not hide that
newline.  This is intentional: if hidden groups are put side-by-side,
there will be problems when moving across them.

I cannot figure out the problem at all: if I call the command
manually, either via eval-expression or via "M-x", then no problems
occur; but if I call the command via a key-binding, it will move to
the wrong place.

So I decided to leave a newline in between the hidden groups.  Indeed
this prevents the wrong motions now.

(ilist-skip-boundary, ilist-skip-properties, ilist-forward-line): Skip
invisible lines by default.  The default behaviour is to skip
invisible lines.  The reason is that the occasion for these commands
without skipping invisible lines is quite rare.