fix previous bugs caused by invisibility
* ilist.el (ilist-string): Now the newlines have the text properties
  inherited from its previous item or group header.

(ilist-belong-to-spec, ilist-invisible-property-p): Helper functions
to determine whether a property is regarded as belonging to a list /
invisible, according to the specifications of the documentation of the
variable `buffer-invisibility-spec'.

(ilist-point-at-eol): The position at the end of the line, possibly
skipping invisible texts.

(ilist-get-property): Get the property from the end of the line,
possibly ignoring invisible texts.

(ilist-get-index, ilist-get-real-index): The index of the line,
possibly ignoring invisible texts.  The latter function is just a
convenient wrapper.

(ilist-get-group): Similar to ilist-get-index.

(ilist-hidden-line-p, ilist-hidden-group-p): Change the name to
latter.  Determine whether the current group is invisible, possibly
ignoring invisible texts.

(ilist-skip-properties, ilist-forward-line,
ilist-forward-group-header): Correctly ignoring invisible texts.