Apply mango to certain Magit heading selections
These faces appear when you select e.g. untracked files, diff
headings, log headings.

It might be too strong to use mango as a selection face here,
especially when the normal ‘region’ face is a subtle grey. However,
the subtle grey background occurs next to the slighly stronger bluish
grey of the ‘highlight’ face used in some Magit contexts, which does
not communicate well that it is an active region/selection to my
eyes. This is probably a wider problem with the ‘highlight’ and
‘region’ faces in general, but it's especially problematic here. So
I've had to resort to mango -- at least it's cohesive with marked
files in dired.