Bump minimal version to 24.3
I am using 'user-error' at one point, that was not available in 24.1.
It is perhaps risky to assume that because nobody has ever complained,
that nobody is using 'shell-command+' on a very old Emacs build, but I
will do so anyway.

If you are affected by this change, send a message to my public inbox:
~pkal/public-inbox@lists.sr.ht