Do not set default-directory, and manually set buffer-file-name
When `eglot-extend-to-xref` is `t`, this allows eglot to manage the
newly opened buffer with the current server instead of starting a new
one.

For clojure-lsp in particular, this does not allow continued xref use to
work. For that, the buffer still needs to be saved to disk I believe.

lsp-mode I think handles this specially for clojure by dumping extracted
files to a temp directory and including that directory in the current
project somehow.