Allow user to relocate extracted files to the previous project.
This allows jarchive to track how the file was visited, and to
potentially move the extracted file into the visting project's directory
where it can be managed by project.el and other libraries like eglot and
cider.

Note that this may leave a hanging eglot process in the temporary
extracted buffer. Setting `eglot-autoshutdown` to 't is recommended to
save resources.

TODO: This process could be made more interactive, like allowing the
user to select the project or source directory with prompts.