Make eglot-extend-to-xref work, breaking support for `zipfile:` uris
This change removes support for the zipfile dependency-scheme in
clojure-lsp, requiring the setting to be initialized always with "jar".

It also requires some changes to eglot to NOT parse jar scheme URIs at
all, leaving the `jar:file://` prefix. Jarchive now only triggers on
file descriptors as this URI. This makes it easy for eglot to forward
references from the jar uri back to clojure-lsp. This allows clojure-lsp
to continue performing analysis when the extracted file is not in the
project.

The following code snippet allows this to work before eglot change are
patched into emacs, (or hacked in with advice if we decide to leave it
out of eglot).

```emacs-lisp
(defun eglot--path-to-uri (path)
  "URIfy PATH."
  (if (equal "jar" (url-type (url-generic-parse-url path)))
      path
    (let ((truepath (file-truename path)))
      (concat "file://"
              ;; Add a leading "/" for local MS Windows-style paths.
              (if (and (eq system-type 'windows-nt)
                       (not (file-remote-p truepath)))
                  "/")
              (url-hexify-string
               ;; Again watch out for trampy paths.
               (directory-file-name (file-local-name truepath))
               eglot--uri-path-allowed-chars)))))

(defun eglot--uri-to-path (uri)
  "Convert URI to file path, helped by `eglot--current-server'."
  (when (keywordp uri) (setq uri (substring (symbol-name uri) 1)))
  (let* ((server (eglot-current-server))
         (remote-prefix (and server (eglot--trampish-p server)))
         (url (url-generic-parse-url uri))
         (retval (if (string= "jar" (url-type url))
                     uri
                   (url-unhex-string (url-filename url))))
         ;; Remove the leading "/" for local MS Windows-style paths.
         (normalized (if (and (not remote-prefix)
                              (eq system-type 'windows-nt)
                              (cl-plusp (length retval)))
                         (substring retval 1)
                       retval)))
    (concat remote-prefix normalized)))
```

This `eglot-server-programs` entry is also used
```
((clojure-mode clojurescript-mode clojurec-mode)
   . ("clojure-lsp" :initializationOptions (:dependency-scheme "jar")))
```

This setting can also be set in
`~/.config/clojure-lsp/config.edn` with `{:dependency-scheme "jar"}`