Do NOT implement file-attributes
At least not in terms of the jar.
Navigating with `find-buffer-visiting` will use file-attributes to try
to id an existing buffer with matching attributes. This means that
navigating to a file in a jar, call it F2, where there exists ANOTHER
open file, called F1, then F1 may be opened instead of F2.

Instead we just return nil, which is what the system file-attributes
implementation will do.