Make persist-defvar work with records
Previously, when persist-defvar received a record for an initial
value, updated values were not persisted.
The behavior of `copy-tree` was changed in Emacs 30. This function will
ensure that persist works correctly for previous Emacs versions.

* persist.el (persist-copy-tree): New function, so that records can
be compared.

* persist.el (persist-symbol): Set default to a copy of initvalue so
when initvalue is a record, the original can be modified without
modifying the default.
* test/persist-tests.el: Test persist-save with a record.