Don't error out on 'nil' alist value
The intention of the DEFAULT argument to 'alist-get' is to return "" instead of
nil.  However, this still returns nil if the key exists but the value is
nil.

This problem wasn't detected previously since the functions
'debbugs-gnu-sort-submitter' and 'debbugs-gnu-sort-title' weren't actually
used until the previous commit.

* debbugs-gnu.el (debbugs-gnu-sort-submitter)
(debbugs-gnu-sort-title): Remove DEFAULT argument of alist-get and
instead wrap call with 'or' to return default.  (Bug#64064)

* debbugs-org.el (debbugs-org-show-reports): If the subject is
empty, use the empty string.