Adding indirect branch / revision filtering
Most instances seem to not support direct branch / revision filtering
in Changes API requests, that is, requests such as

<host>/api/v2/changes?revision=<revision-id>

timeout or return an error code.

In such cases, we issue a more general API request, and filter the
response by revision:

<host>/api/v2/changes?limit=<limit>&order=-changeid

We add a new custom var `buildbot-api-changes-direct-filter' to
control the relevant behaviour.

The mariadb instance seems to be the only instance that work better
with direct filtering, whereas python and buildbot seems to work with
both, and the remaining instances only supporting indirect filtering.

We also add a few more custom vars for limiting in API requests, and
fix some regressions introduced in the previous commit, about setting
the host.